requirements.txt
pandas
numpy
scipy
plotly
playwright
python-dotenv
